requirements.txt
google-generativeai
python-dotenv
pydantic
matplotlib
